backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
replicate>=0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
httpx==0.27.2
celery[redis]==5.4.0
redis==5.2.0
cloudinary>=1.36.0
pillow==11.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
anthropic>=0.40.0
fal-client>=0.10.0
aiosmtplib>=3.0.0
